requirements.txt
python-gitlab>=4.0.0
python-dotenv>=1.0.0
click>=8.0.0
plyer>=2.0.0
aiohttp>=3.8.0